refactor: TestController의 메서드들을 AuthController로 이동
